Moved helpers from transactions lib to separate file
